requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
fpdf==1.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
mysql-connector-python==9.1.0
numpy==2.1.3
packaging==24.1
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
